refactor(test): reorganize test directory structure
- setup-files/ → setup/
- helpers/ → factories/ and mocks/
- factories/: test data generation factories
- mocks/: vi.fn()-based mock objects (absorbs port-mocks)
